requirements.txt
# ──────────────────────────────────────────────────────────────────
#  Person Re-ID with DeepSORT + DukeMTMC-ReID
#  Requirements
# ──────────────────────────────────────────────────────────────────

# ── Deep Learning ─────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0

# ── YOLOv8 Object Detection ───────────────────────────────────────
ultralytics>=8.0.0

# ── DeepSORT Tracker ──────────────────────────────────────────────
deep-sort-realtime>=1.3.2

# ── Computer Vision ───────────────────────────────────────────────
opencv-python>=4.7.0
Pillow>=9.0.0

# ── Numerical / Scientific ────────────────────────────────────────
numpy>=1.24.0
scipy>=1.10.0

# ── Machine Learning / Metrics ────────────────────────────────────
scikit-learn>=1.2.0

# ── Visualisation ─────────────────────────────────────────────────
matplotlib>=3.6.0
seaborn>=0.12.0

# ── Data / Tabular ────────────────────────────────────────────────
pandas>=1.5.0

# ── Training Utilities ────────────────────────────────────────────
tqdm>=4.64.0
tensorboard>=2.12.0
gdown>=4.7.1

# ── Optional: OSNet / torchreid backbone (Technique 5) ───────────
 torchreid  # uncomment to enable OSNet support
